将商品的最高，最低质保年限改为可设为小数的

+++ dangjia-group/dangjia-goods/dangjia-api-goods/src/main/java/com/dangjia/acg/dto/product/BasicsProductDTO.java
@@ -66,9 +66,9 @@ public class BasicsProductDTO extends GoodsBaseEntity {
     private String isInflueWarrantyPeriod;//是否影响质保年限（1是，0否）
 
 
-    private Integer maxWarrantyPeriodYear;//最高质保年限
+    private Double maxWarrantyPeriodYear;//最高质保年限
 
-    private Integer minWarrantyPeriodYear;//最低质保年限
+    private Double minWarrantyPeriodYear;//最低质保年限
 
     private String marketingName;//营销名称
